Function kaleidoscope_ast::node::reify_node

pub fn reify_node<N>(node: Box<dyn Node>) -> Option<Box<N>> where
    N: Node + NodeType, 

Convert a node as a trait object and convert it into a concrete node with a known type.

Example

use kaleidoscope_ast::{
    node::{reify_node, Node},
    nodes::IntegerNode
};

let unknown: Box<dyn Node> = Box::new(IntegerNode::new(10));
let resolved: Box<IntegerNode> = reify_node(unknown).unwrap();
assert_eq!(resolved.get_value(), 10);