Trait kaleidoscope_ast::node::NodeType

pub trait NodeType: Sized + 'static {
    fn node_id() -> NodeId { ... }
}

Types which act as nodes in an AST of a Kaleidoscope programme. This type is separated from Node so that the Rust compiler can create trait objects from Node and ExprNode.

Provided methods

fn node_id() -> NodeId

Get the NodeId of a node type.